requirements.txt
Adafruit-Blinka==8.62.0
adafruit-circuitpython-busdevice==5.2.13
adafruit-circuitpython-connectionmanager==3.1.5
adafruit-circuitpython-requests==4.1.13
adafruit-circuitpython-tsl2591==1.4.4
adafruit-circuitpython-typing==1.12.1
Adafruit-DHT==1.4.0
Adafruit-PlatformDetect==3.81.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
blinker==1.9.0
board==1.0
click==8.2.1
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pyftdi==0.56.0
pyserial==3.5
pyusb==1.3.1
RPi.GPIO==0.7.1
schedule==1.2.2
sysv-ipc==1.1.0
typing_extensions==4.14.1
Werkzeug==3.1.3
